requirements/base.txt
asgiref==3.3.4
blessings==1.7
bpython==0.21
certifi==2020.12.5
chardet==4.0.0
curtsies==0.3.5
cwcwidth==0.1.4
dj-stripe==2.5.0
Django==3.2
django-ckeditor==6.0.0
django-colorfield==0.4.1
django-cors-headers==3.7.0
django-filter==2.4.0
django-js-asset==1.2.2
django-solo==1.1.5
djangorestframework==3.12.4
greenlet==1.0.0
gunicorn==20.1.0
idna==2.10
jsonfield==3.1.0
Markdown==3.3.4
martor==1.6.3
Pillow==8.2.0
psycopg2-binary==2.8.6
Pygments==2.8.1
pytz==2021.1
pyxdg==0.27
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
stripe==2.58.0
urllib3==1.26.4
